requirements.txt
comtradeapicall
pandas
pandas
openpyxl
psycopg2-binary
matplotlib
Pillow
pytesseract
tqdm
python-dotenv
google-generativeai
chromadb
tabulate
pdf2image
pypdf
pdfplumber
langchain-core
langchain-chroma
langchain-google-genai
langchain-huggingface
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
sentence-transformers
unstructured[all-docs]
faiss-cpu
PyMuPDF
fastapi
uvicorn[standard]
pydantic